Move generic components to another folder
