+++ src/components/generic/SvgImage.tsx
